+++ src/main/java/ultigreen/demo/controller/APIReusableItems.java
@@ -1,7 +1,5 @@
 package ultigreen.demo.controller;
 
-import java.util.List;
-
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 
@@ -14,11 +12,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import ultigreen.demo.ReusablesService;
-import ultigreen.demo.domain.AppUser;
+import ultigreen.demo.services.ReusablesService;
 import ultigreen.demo.domain.ReusableQuestion;
 
 @RestController
